Move step function helpers to separate module
